Add files via upload
Hello World! This was a project that was 10 hours in the making. It was a coding challenge given by a company that is interested in hiring me. I learned so much during this time. I was frustrated, did more tests than I can count on my hand but it was all worth it watching this little guy move. I hope I get the job, but if not I still learned a few new things and had the most fun along the way. console.log("Here We Go Again");